budget: price the transmit half of a received segment
The receive budget stopped at the driver: reap/stuff/post priced ONE
tx_send and ack priced the wire, but nothing measured the stack work
between a segment reaching its socket and the ACK reaching the driver,
so a TX cut could not be shown to have cut anything.  One leg, the
existing mechanism: socket_enter arms xmit_at, tx_send consumes it once,
deliver clears a stale one.  nrb_Xmit is APPENDED, offsets hold, so
AMI_NETSTATUS_VERSION stays 11 and BSD_LIB_REVISION 7.  sana2_rx.c's
sign-compare came with it: the RXPROBE arm did not build at all.

+++ src/sana2/sana2_rx.c
@@ -123,7 +123,7 @@ VOID ami_sana2_rxprobe_deliver(AmiSana2If *iface, const UCHAR *frame,
     ihl   = (UWORD)((ip[0] & 0x0F) * 4);
     total = (ULONG)ami_rxprobe_be16(&ip[2]);
 
-    if (ihl < 20 || total < ihl + 20 ||
+    if (ihl < 20 || total < (ULONG)ihl + 20UL ||
         total > length - AMI_ETH_HEADER_SIZE)
         return;
